fix(serve): emit all stream_chunk deltas to fix concurrent tool-call streaming
