+++ lmdeploy/serve/parsers/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 # registers ResponseParser 'gpt-oss', None if openai_harmony unavailable
 from .gpt_oss_response_parser import GptOssResponseParser
-from .response_parser import ResponseParser, ResponseParserManager, first_stream_delta
+from .response_parser import ResponseParser, ResponseParserManager
 
-__all__ = ['ResponseParser', 'ResponseParserManager', 'GptOssResponseParser', 'first_stream_delta']
+__all__ = ['ResponseParser', 'ResponseParserManager', 'GptOssResponseParser']
