+++ tests/test_lmdeploy/serve/parsers/test_glm47_parser.py
@@ -3,9 +3,11 @@ import json
 import pytest
 
 from lmdeploy.serve.openai.protocol import ChatCompletionRequest
-from lmdeploy.serve.parsers import ResponseParserManager, first_stream_delta
+from lmdeploy.serve.parsers import ResponseParserManager
 from lmdeploy.serve.parsers.reasoning_parser import ReasoningParserManager
 from lmdeploy.serve.parsers.tool_parser import Glm47ToolParser, ToolParserManager
+
+from .helpers import first_stream_delta
 
 MODEL_ID = 'zai-org/GLM-4.7'
